Time: 259 ms (42.27%) | Memory: 0B (100.00%) - LeetSync

+++ 2087-confirmation-rate/confirmation-rate.sql
@@ -1,5 +1,9 @@
 -- Write your PostgreSQL query statement below
-select s.user_id, round((sum(case when c.action = 'confirmed' then 1 else 0 end) * 1.0 / count(*))::numeric, 2) confirmation_rate
+select s.user_id, 
+round(
+    sum(case when c.action = 'confirmed' then 1 else 0 end) * 1.0 / count(*)
+    , 2
+    ) confirmation_rate
 from signups s
 left join confirmations c
 on s.user_id = c.user_id
